wip: consolidate
