+++ src/routing/index.ts
@@ -1,5 +1,5 @@
-export type { RouteSource } from '@/routing/registration/source/route-source';
+export type { RouteSource } from '@/routing/declaration/route-source.type';
 export { Route } from './declaration/route';
 export { createPathFor } from './path/create-path-for';
 export { RouteGroup } from './declaration/route-group';
-export { Any, Delete, Get, Head, Options, Patch, Post, Put } from '@/routing/registration/verb/http-verb-helpers';
+export { Any, Delete, Get, Head, Options, Patch, Post, Put } from '@/routing/declaration/http-verb-helpers';
