+++ src/routing/path/create-path-for.test.ts
@@ -1,4 +1,4 @@
-import { Get } from '@/routing/registration/verb/http-verb-helpers';
+import { Get } from '@/routing/declaration/http-verb-helpers';
 import { describe, expect, test, vi } from 'vitest';
 import { RouteGroup } from '../declaration/route-group';
 import { createPathFor } from './create-path-for';
