+++ src/routing/registration/source/create-path-catalog.test.ts
@@ -1,6 +1,6 @@
+import { Get } from '@/routing/declaration/http-verb-helpers';
 import { Route } from '@/routing/declaration/route';
 import { RouteGroup } from '@/routing/declaration/route-group';
-import { Get } from '@/routing/registration/verb/http-verb-helpers';
 import { describe, expect, test, vi } from 'vitest';
 import { createPathCatalog } from './create-path-catalog';
