+++ src/routing/registration/source/normalize-route-sources.ts
@@ -2,7 +2,7 @@ import type { NormalizedRouteProps } from '@/routing/declaration/normalized-rout
 import type { RouteGroupDefinition } from '@/routing/declaration/route-group';
 import { mergeRouteOptions } from '@/routing/registration/source/resolve-route-options';
 import type { Request } from 'koa';
-import type { RouteSource } from './route-source';
+import type { RouteSource } from '../../declaration/route-source.type';
 
 interface NormalizationContext<TRequest extends Request> {
   prefix: string;
